Revert "기존 의도에 걸맞는 아키텍처로 재구조"
This reverts commit a92ae83739cccdf946df7d0b482e908d94cc1b44.
