+++ src/main/java/com/sparta/miniproject/common/config/chain/container/CsrfContainer.java
@@ -1,15 +1,16 @@
-package com.sparta.miniproject.common.config.container;
+package com.sparta.miniproject.common.config.chain.container;
 
 import com.sparta.miniproject.common.security.CustomizerContainer;
 import org.springframework.security.config.Customizer;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
+import org.springframework.security.config.annotation.web.configurers.CsrfConfigurer;
 import org.springframework.stereotype.Component;
 
 import java.util.Collection;
 
 @Component
-public class FilterChainContainer extends CustomizerContainer<HttpSecurity> {
-    public FilterChainContainer(Collection<Customizer<HttpSecurity>> elements) {
+public class CsrfContainer extends CustomizerContainer<CsrfConfigurer<HttpSecurity>> {
+    public CsrfContainer(Collection<Customizer<CsrfConfigurer<HttpSecurity>>> elements) {
         super(elements);
     }
 }
